doh, fixed a major accuracy issue switching form int -> double for internal storage of values
git-svn-id: https://pbtech-vc.med.cornell.edu/public/svn/icb/trunk/icb-commons/icb-utils@17291 60cd2ccc-2b05-0410-9566-da273a940bff

+++ src/edu/cornell/med/icb/stat/LinearRegression.java
@@ -28,16 +28,16 @@ public class LinearRegression {
     private int numberDataPoints;
 
     /** The sum of all data point x*x values. */
-    private int sumxx;
+    private double sumxx;
 
     /** The sum of all data point x*y values. */
-    private int sumxy;
+    private double sumxy;
 
     /** The sum of all data point x values. */
-    private int sumx;
+    private double sumx;
 
     /** The sum of all data point y values. */
-    private int sumy;
+    private double sumy;
 
     /** The x-intercept as calculated in regress(). */
     private double xIntercept;
@@ -60,7 +60,7 @@ public class LinearRegression {
      * @param x the x of the point being added
      * @param y the y of the point being added
      */
-    public void addDataPoint(final int x, final int y) {
+    public void addDataPoint(final double x, final double y) {
         numberDataPoints++;
         sumx += x;
         sumy += y;
